requirements.txt
langchain
groq
langchain-groq
pydantic
typing
python-dotenv
field